features/compress: Compression/DeCompression translator
This is the on-wire compression translator. Data is compressed
before it is sent out to the network and decompressed on the other
side. What is saved is the amount of data thats sent over the wire.

Keep reading...

This translator has two modes of operation: CLIENT and SERVER

The two operation modes behave differently based on the type of
fop in progress (readv()/writev()):
 * Operation mode 'SERVER' compresses/decompresses data in the
   readv() and writev() call on the server respectively.
 * Operation mode 'CLIENT' compresses/decompresses data in the
   writev() and readv() call on the client respectively.

(De)Compression is done using Zlib library (inflate/deflate
routines respectively). Data transfer format is similar to the
gzip except minus the header (ie. <compressed data> + trailer).
Trailer is needed as it contains the CRC and length of the
original (uncompressed) data used for validating inflated
data.

A debug option is provided which when turned on (from cli), dumps
the compressed data (including the gzip header) to a file. This
can be used while writing unit test cases to check the validity
of the compressed data uisng standard unix tools (gzip).

HOW TO USE
----------

Ex : gluster volume set <vol_name> compress on         # turn on compression
     gluster volume set <vol_name> compress.buffer-size 32768
     gluster volume set <vol_name> compress.debug on   # turn on debugging

Credit goes to Prashanth Pai (nullpai@gmail.com) in extending the
translator for writev() fop, adding CLI support to turn on/off
compression, unit test cases and many other code improvements.

Signed-off-by: Venky Shankar <vshankar@redhat.com>